#2133 moved around components
